[Silver III] Title: 주유소, Time: 412 ms, Memory: 40520 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -1,31 +1,31 @@
 import java.io.BufferedReader;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.Arrays;
 
 public class Main {
     static int N; // N(도시 개수)
-    static int[] distance;
-    static int[] price;
+    static long[] distance;
+    static long[] price;
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         N = Integer.parseInt(br.readLine());
         distance = Arrays.stream(br.readLine().split(" "))
-                .mapToInt(Integer::parseInt)
+                .mapToLong(Long::parseLong)
                 .toArray();
         price = Arrays.stream(br.readLine().split(" "))
-                .mapToInt(Integer::parseInt)
+                .mapToLong(Long::parseLong)
                 .toArray();
 
-        int prevPrice = price[0];
-        int answer = 0;
+        long prevPrice = price[0];
+        long answer = 0;
         for (int i = 1; i < price.length; i++) {
             answer += prevPrice * distance[i - 1];
             prevPrice = Math.min(prevPrice, price[i]);
         }
+        
         System.out.println(answer);
     }
 }
